+++ lib/shim/datastore-shim.js
@@ -15,6 +15,7 @@ const Shim = require('./shim')
 const urltils = require('../util/urltils')
 const util = require('util')
 const specs = require('./specs')
+const recordOperationMetrics = require('../../lib/metrics/recorders/record-operation-metrics')
 const { DatastoreParameters } = specs.params
 
 /**
@@ -284,7 +285,7 @@ function recordOperation(nodule, properties, opSpec) {
       if (!segDesc?.name?.startsWith(shim._metrics.OPERATION)) {
         segDesc.name = shim._metrics.OPERATION + segDesc.name
       }
-      segDesc.recorder = _recordOperationMetrics.bind(shim)
+      segDesc.recorder = recordOperationMetrics.bind(shim)
     }
 
     return segDesc
@@ -609,58 +610,6 @@ function _getSpec({ spec, shim, fn, fnName, args }) {
 function _recordQueryMetrics(parsed, segment, scope) {
   if (segment) {
     parsed.recordMetrics(segment, scope)
-  }
-}
-
-/**
- * Records all the metrics required for database operations.
- *
- * - `_recordOperationMetrics(segment [, scope])`
- *
- * @private
- * @this DatastoreShim
- * @implements {MetricFunction}
- * @param {TraceSegment}  segment - The segment being recorded.
- * @param {string}        [scope] - The scope of the segment.
- * @see DatastoreShim#recordOperation
- * @see MetricFunction
- */
-function _recordOperationMetrics(segment, scope) {
-  if (!segment) {
-    return
-  }
-
-  const duration = segment.getDurationInMillis()
-  const exclusive = segment.getExclusiveDurationInMillis()
-  const transaction = segment.transaction
-  const type = transaction.isWeb() ? 'allWeb' : 'allOther'
-  const operation = segment.name
-
-  if (scope) {
-    transaction.measure(operation, scope, duration, exclusive)
-  }
-
-  transaction.measure(operation, null, duration, exclusive)
-  transaction.measure(metrics.DB.PREFIX + type, null, duration, exclusive)
-  transaction.measure(metrics.DB.ALL, null, duration, exclusive)
-  transaction.measure(this._metrics.ALL, null, duration, exclusive)
-  transaction.measure(
-    metrics.DB.PREFIX + this._metrics.PREFIX + '/' + type,
-    null,
-    duration,
-    exclusive
-  )
-
-  const attributes = segment.getAttributes()
-  if (attributes.host && attributes.port_path_or_id) {
-    const instanceName = [
-      metrics.DB.INSTANCE,
-      this._metrics.PREFIX,
-      attributes.host,
-      attributes.port_path_or_id
-    ].join('/')
-
-    transaction.measure(instanceName, null, duration, exclusive)
   }
 }
