+++ lib/shim/webframework-shim/middleware.js
@@ -13,6 +13,7 @@ const {
 } = require('./common')
 const { assignCLMSymbol } = require('../../util/code-level-metrics')
 const { RecorderSpec } = require('../specs')
+const makeMiddlewareRecorder = require('../../metrics/recorders/middleware-metric-recorder')
 
 const MIDDLEWARE_TYPE_DETAILS = {
   APPLICATION: { name: 'Mounted App: ', path: true, record: false },
@@ -88,7 +89,7 @@ function constructRecorder({ txInfo, typeDetails, shim, metricName }) {
   let recorder = null
   if (typeDetails.record) {
     const stackPath = txInfo.transaction.nameState.getPath() || ''
-    recorder = _makeMiddlewareRecorder(shim, metricName + '/' + stackPath)
+    recorder = makeMiddlewareRecorder(shim, metricName + '/' + stackPath)
   }
   return recorder
 }
@@ -323,27 +324,6 @@ module.exports._recordMiddleware = function _recordMiddleware(shim, middleware, 
       ? middlewareWithPromiseRecorder({ spec, typeDetails, metricName, isErrorWare })
       : middlewareWithCallbackRecorder({ spec, typeDetails, metricName, isErrorWare })
   )
-}
-
-/**
- * Creates a recorder for middleware metrics.
- *
- * @private
- * @param {object} _shim instance of shim
- * @param {string} metricName name of metric
- * @returns {Function} recorder for middleware
- */
-function _makeMiddlewareRecorder(_shim, metricName) {
-  return function middlewareMetricRecorder(segment, scope) {
-    const duration = segment.getDurationInMillis()
-    const exclusive = segment.getExclusiveDurationInMillis()
-    const transaction = segment.transaction
-
-    if (scope) {
-      transaction.measure(metricName, scope, duration, exclusive)
-    }
-    transaction.measure(metricName, null, duration, exclusive)
-  }
 }
 
 /**
